+++ lib/main/storages/prefs_storage.dart
@@ -1,0 +1,41 @@
+import 'package:shared_preferences/shared_preferences.dart';
+
+class PrefsStorage {
+  /* Variables */
+
+  SharedPreferences? prefs;
+
+  /* Methods */
+
+  Future<void> init() async {
+    if (prefs != null) {
+      return;
+    }
+
+    prefs = await SharedPreferences.getInstance();
+  }
+
+  int readInt({
+    required String key,
+    int fallback = 0,
+  }) {
+    return prefs?.getInt(key) ?? fallback;
+  }
+
+  Future<void> writeInt({
+    required String key,
+    required int value,
+  }) async {
+    await prefs?.setInt(key, value);
+  }
+
+  Future<void> delete({
+    required String key,
+  }) async {
+    await prefs?.remove(key);
+  }
+
+  Future<void> deleteAll() async {
+    await prefs?.clear();
+  }
+}